+++ tests/renderers/TypeScript.test.js
@@ -257,6 +257,18 @@ describe('TypeScriptRenderer', () => {
     });
   });
 
+  describe('renderSchema()', () => {
+    it('renders with generics', () => {
+      renderer.options.schemaGenerics = true;
+
+      expect(
+        renderer.renderSchema('QuxSchema', [], {
+          resourceName: 'quxs',
+        }),
+      ).toBe("export const quxSchema = new Schema<QuxInterface>('quxs');");
+    });
+  });
+
   describe('renderString()', () => {
     it('renders nullable', () => {
       expect(renderer.renderString(new StringDefinition(options, 'foo'))).toBe('string | null');